PRICEAI_API/requirements.txt
# Core Framework
fastapi
uvicorn
supabase

# Configuration & Environment
python-dotenv

# Utilities & Formatting  
rich

# HTTP Clients
aiohttp
requests

